1.4.99 - Major change to the code to add experiments, a feature that enables multiple sets of parameters to be run simultaneously, while other variables are changed automatically. This change is deeply embedded in the support code but hardly affects the simulation code.
*** Spectral analysis stuff is currently broken, hence the separate branch ***

Another major issue was resolved on this revision that was only discovered with the most recent testing. gids are now being cleared prior to creating a new simulation in a run, so that they can be reassigned without confusing neuron with multiple, incompatible assignments. This change should be pushed mainline as is necessary.

Other minor changes were made.