1.7.37 - Fixed specmax routine in cli that gives max of spectral output (requires data to be present), along with frequency/time at peak. Calculates and reports suggested interval on which to do stim.
Also there is a new input method for arange for the param files:
some_param: A[start, end, step]

This does the interval from [start, end), so it is not inclusive at the end. There is a chance the addition of arange breaks something downstream, but I haven't figure out what/where/why etc. yet.